Move APIs back to internal headers
