+++ config.m4
@@ -115,14 +115,6 @@ if test "$PHP_GRPC" != "no"; then
     src/core/ext/filters/message_size/message_size_filter.cc \
     src/core/ext/filters/workarounds/workaround_cronet_compression_filter.cc \
     src/core/ext/filters/workarounds/workaround_utils.cc \
-    src/core/ext/transport/binder/server/binder_server.cc \
-    src/core/ext/transport/binder/transport/binder_transport.cc \
-    src/core/ext/transport/binder/utils/transport_stream_receiver_impl.cc \
-    src/core/ext/transport/binder/wire_format/binder_android.cc \
-    src/core/ext/transport/binder/wire_format/binder_constants.cc \
-    src/core/ext/transport/binder/wire_format/transaction.cc \
-    src/core/ext/transport/binder/wire_format/wire_reader_impl.cc \
-    src/core/ext/transport/binder/wire_format/wire_writer.cc \
     src/core/ext/transport/chttp2/alpn/alpn.cc \
     src/core/ext/transport/chttp2/client/authority.cc \
     src/core/ext/transport/chttp2/client/chttp2_connector.cc \
@@ -1094,10 +1086,6 @@ if test "$PHP_GRPC" != "no"; then
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/filters/max_age)
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/filters/message_size)
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/filters/workarounds)
-  PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/binder/server)
-  PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/binder/transport)
-  PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/binder/utils)
-  PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/binder/wire_format)
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/chttp2/alpn)
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/chttp2/client)
   PHP_ADD_BUILD_DIR($ext_builddir/src/core/ext/transport/chttp2/client/insecure)
