+++ gRPC-C++.podspec
@@ -265,18 +265,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/message_size/message_size_filter.h',
                       'src/core/ext/filters/workarounds/workaround_cronet_compression_filter.h',
                       'src/core/ext/filters/workarounds/workaround_utils.h',
-                      'src/core/ext/transport/binder/server/binder_server.h',
-                      'src/core/ext/transport/binder/transport/binder_stream.h',
-                      'src/core/ext/transport/binder/transport/binder_transport.h',
-                      'src/core/ext/transport/binder/utils/transport_stream_receiver.h',
-                      'src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h',
-                      'src/core/ext/transport/binder/wire_format/binder.h',
-                      'src/core/ext/transport/binder/wire_format/binder_android.h',
-                      'src/core/ext/transport/binder/wire_format/binder_constants.h',
-                      'src/core/ext/transport/binder/wire_format/transaction.h',
-                      'src/core/ext/transport/binder/wire_format/wire_reader.h',
-                      'src/core/ext/transport/binder/wire_format/wire_reader_impl.h',
-                      'src/core/ext/transport/binder/wire_format/wire_writer.h',
                       'src/core/ext/transport/chttp2/alpn/alpn.h',
                       'src/core/ext/transport/chttp2/client/authority.h',
                       'src/core/ext/transport/chttp2/client/chttp2_connector.h',
@@ -811,7 +799,6 @@ Pod::Spec.new do |s|
                       'src/cpp/common/validate_service_config.cc',
                       'src/cpp/common/version_cc.cc',
                       'src/cpp/server/async_generic_service.cc',
-                      'src/cpp/server/binder_server_credentials.cc',
                       'src/cpp/server/channel_argument_option.cc',
                       'src/cpp/server/create_default_thread_pool.cc',
                       'src/cpp/server/dynamic_thread_pool.cc',
@@ -945,18 +932,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/filters/message_size/message_size_filter.h',
                               'src/core/ext/filters/workarounds/workaround_cronet_compression_filter.h',
                               'src/core/ext/filters/workarounds/workaround_utils.h',
-                              'src/core/ext/transport/binder/server/binder_server.h',
-                              'src/core/ext/transport/binder/transport/binder_stream.h',
-                              'src/core/ext/transport/binder/transport/binder_transport.h',
-                              'src/core/ext/transport/binder/utils/transport_stream_receiver.h',
-                              'src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h',
-                              'src/core/ext/transport/binder/wire_format/binder.h',
-                              'src/core/ext/transport/binder/wire_format/binder_android.h',
-                              'src/core/ext/transport/binder/wire_format/binder_constants.h',
-                              'src/core/ext/transport/binder/wire_format/transaction.h',
-                              'src/core/ext/transport/binder/wire_format/wire_reader.h',
-                              'src/core/ext/transport/binder/wire_format/wire_reader_impl.h',
-                              'src/core/ext/transport/binder/wire_format/wire_writer.h',
                               'src/core/ext/transport/chttp2/alpn/alpn.h',
                               'src/core/ext/transport/chttp2/client/authority.h',
                               'src/core/ext/transport/chttp2/client/chttp2_connector.h',
