+++ grpc.gemspec
@@ -244,26 +244,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/workarounds/workaround_cronet_compression_filter.h )
   s.files += %w( src/core/ext/filters/workarounds/workaround_utils.cc )
   s.files += %w( src/core/ext/filters/workarounds/workaround_utils.h )
-  s.files += %w( src/core/ext/transport/binder/server/binder_server.cc )
-  s.files += %w( src/core/ext/transport/binder/server/binder_server.h )
-  s.files += %w( src/core/ext/transport/binder/transport/binder_stream.h )
-  s.files += %w( src/core/ext/transport/binder/transport/binder_transport.cc )
-  s.files += %w( src/core/ext/transport/binder/transport/binder_transport.h )
-  s.files += %w( src/core/ext/transport/binder/utils/transport_stream_receiver.h )
-  s.files += %w( src/core/ext/transport/binder/utils/transport_stream_receiver_impl.cc )
-  s.files += %w( src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/binder.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/binder_android.cc )
-  s.files += %w( src/core/ext/transport/binder/wire_format/binder_android.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/binder_constants.cc )
-  s.files += %w( src/core/ext/transport/binder/wire_format/binder_constants.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/transaction.cc )
-  s.files += %w( src/core/ext/transport/binder/wire_format/transaction.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/wire_reader.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/wire_reader_impl.cc )
-  s.files += %w( src/core/ext/transport/binder/wire_format/wire_reader_impl.h )
-  s.files += %w( src/core/ext/transport/binder/wire_format/wire_writer.cc )
-  s.files += %w( src/core/ext/transport/binder/wire_format/wire_writer.h )
   s.files += %w( src/core/ext/transport/chttp2/alpn/alpn.cc )
   s.files += %w( src/core/ext/transport/chttp2/alpn/alpn.h )
   s.files += %w( src/core/ext/transport/chttp2/client/authority.cc )
