+++ include/grpcpp/security/server_credentials.h
@@ -64,13 +64,6 @@ namespace experimental {
 /// Builds Xds ServerCredentials given fallback credentials
 std::shared_ptr<ServerCredentials> XdsServerCredentials(
     const std::shared_ptr<ServerCredentials>& fallback_credentials);
-
-/// Builds Binder ServerCredentials.
-///
-/// Calling \a ServerBuilder::AddListeningPort() with Binder ServerCredentials
-/// in a non-Android environment will make the subsequent call to
-/// \a ServerBuilder::BuildAndStart() returns a null pointer.
-std::shared_ptr<ServerCredentials> BinderServerCredentials();
 }  // namespace experimental
 
 /// Wrapper around \a grpc_server_credentials, a way to authenticate a server.
