+++ package.xml
@@ -224,26 +224,6 @@
     <file baseinstalldir="/" name="src/core/ext/filters/workarounds/workaround_cronet_compression_filter.h" role="src" />
     <file baseinstalldir="/" name="src/core/ext/filters/workarounds/workaround_utils.cc" role="src" />
     <file baseinstalldir="/" name="src/core/ext/filters/workarounds/workaround_utils.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/server/binder_server.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/server/binder_server.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/transport/binder_stream.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/transport/binder_transport.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/transport/binder_transport.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/utils/transport_stream_receiver.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/utils/transport_stream_receiver_impl.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/binder.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/binder_android.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/binder_android.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/binder_constants.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/binder_constants.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/transaction.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/transaction.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/wire_reader.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/wire_reader_impl.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/wire_reader_impl.h" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/wire_writer.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/ext/transport/binder/wire_format/wire_writer.h" role="src" />
     <file baseinstalldir="/" name="src/core/ext/transport/chttp2/alpn/alpn.cc" role="src" />
     <file baseinstalldir="/" name="src/core/ext/transport/chttp2/alpn/alpn.h" role="src" />
     <file baseinstalldir="/" name="src/core/ext/transport/chttp2/client/authority.cc" role="src" />
