+++ src/ruby/ext/grpc/rb_grpc_imports.generated.c
@@ -84,7 +84,6 @@ grpc_server_config_fetcher_xds_create_type grpc_server_config_fetcher_xds_create
 grpc_server_config_fetcher_destroy_type grpc_server_config_fetcher_destroy_import;
 grpc_server_set_config_fetcher_type grpc_server_set_config_fetcher_import;
 grpc_server_add_insecure_http2_port_type grpc_server_add_insecure_http2_port_import;
-grpc_server_add_binder_port_type grpc_server_add_binder_port_import;
 grpc_server_start_type grpc_server_start_import;
 grpc_server_shutdown_and_notify_type grpc_server_shutdown_and_notify_import;
 grpc_server_cancel_all_calls_type grpc_server_cancel_all_calls_import;
@@ -373,7 +372,6 @@ void grpc_rb_load_imports(HMODULE library) {
   grpc_server_config_fetcher_destroy_import = (grpc_server_config_fetcher_destroy_type) GetProcAddress(library, "grpc_server_config_fetcher_destroy");
   grpc_server_set_config_fetcher_import = (grpc_server_set_config_fetcher_type) GetProcAddress(library, "grpc_server_set_config_fetcher");
   grpc_server_add_insecure_http2_port_import = (grpc_server_add_insecure_http2_port_type) GetProcAddress(library, "grpc_server_add_insecure_http2_port");
-  grpc_server_add_binder_port_import = (grpc_server_add_binder_port_type) GetProcAddress(library, "grpc_server_add_binder_port");
   grpc_server_start_import = (grpc_server_start_type) GetProcAddress(library, "grpc_server_start");
   grpc_server_shutdown_and_notify_import = (grpc_server_shutdown_and_notify_type) GetProcAddress(library, "grpc_server_shutdown_and_notify");
   grpc_server_cancel_all_calls_import = (grpc_server_cancel_all_calls_type) GetProcAddress(library, "grpc_server_cancel_all_calls");
