+++ test/core/surface/public_headers_must_be_c89.c
@@ -130,7 +130,6 @@ int main(int argc, char **argv) {
   printf("%lx", (unsigned long) grpc_server_config_fetcher_destroy);
   printf("%lx", (unsigned long) grpc_server_set_config_fetcher);
   printf("%lx", (unsigned long) grpc_server_add_insecure_http2_port);
-  printf("%lx", (unsigned long) grpc_server_add_binder_port);
   printf("%lx", (unsigned long) grpc_server_start);
   printf("%lx", (unsigned long) grpc_server_shutdown_and_notify);
   printf("%lx", (unsigned long) grpc_server_cancel_all_calls);
