+++ tools/doxygen/Doxyfile.c++.internal
@@ -1181,26 +1181,6 @@ src/core/ext/filters/workarounds/workaround_cronet_compression_filter.cc \
 src/core/ext/filters/workarounds/workaround_cronet_compression_filter.h \
 src/core/ext/filters/workarounds/workaround_utils.cc \
 src/core/ext/filters/workarounds/workaround_utils.h \
-src/core/ext/transport/binder/server/binder_server.cc \
-src/core/ext/transport/binder/server/binder_server.h \
-src/core/ext/transport/binder/transport/binder_stream.h \
-src/core/ext/transport/binder/transport/binder_transport.cc \
-src/core/ext/transport/binder/transport/binder_transport.h \
-src/core/ext/transport/binder/utils/transport_stream_receiver.h \
-src/core/ext/transport/binder/utils/transport_stream_receiver_impl.cc \
-src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h \
-src/core/ext/transport/binder/wire_format/binder.h \
-src/core/ext/transport/binder/wire_format/binder_android.cc \
-src/core/ext/transport/binder/wire_format/binder_android.h \
-src/core/ext/transport/binder/wire_format/binder_constants.cc \
-src/core/ext/transport/binder/wire_format/binder_constants.h \
-src/core/ext/transport/binder/wire_format/transaction.cc \
-src/core/ext/transport/binder/wire_format/transaction.h \
-src/core/ext/transport/binder/wire_format/wire_reader.h \
-src/core/ext/transport/binder/wire_format/wire_reader_impl.cc \
-src/core/ext/transport/binder/wire_format/wire_reader_impl.h \
-src/core/ext/transport/binder/wire_format/wire_writer.cc \
-src/core/ext/transport/binder/wire_format/wire_writer.h \
 src/core/ext/transport/chttp2/alpn/alpn.cc \
 src/core/ext/transport/chttp2/alpn/alpn.h \
 src/core/ext/transport/chttp2/client/authority.cc \
@@ -2279,7 +2259,6 @@ src/cpp/common/tls_credentials_options_util.h \
 src/cpp/common/validate_service_config.cc \
 src/cpp/common/version_cc.cc \
 src/cpp/server/async_generic_service.cc \
-src/cpp/server/binder_server_credentials.cc \
 src/cpp/server/channel_argument_option.cc \
 src/cpp/server/create_default_thread_pool.cc \
 src/cpp/server/dynamic_thread_pool.cc \
