+++ tools/doxygen/Doxyfile.core.internal
@@ -1010,26 +1010,6 @@ src/core/ext/filters/workarounds/workaround_utils.cc \
 src/core/ext/filters/workarounds/workaround_utils.h \
 src/core/ext/transport/README.md \
 src/core/ext/transport/binder/README.md \
-src/core/ext/transport/binder/server/binder_server.cc \
-src/core/ext/transport/binder/server/binder_server.h \
-src/core/ext/transport/binder/transport/binder_stream.h \
-src/core/ext/transport/binder/transport/binder_transport.cc \
-src/core/ext/transport/binder/transport/binder_transport.h \
-src/core/ext/transport/binder/utils/transport_stream_receiver.h \
-src/core/ext/transport/binder/utils/transport_stream_receiver_impl.cc \
-src/core/ext/transport/binder/utils/transport_stream_receiver_impl.h \
-src/core/ext/transport/binder/wire_format/binder.h \
-src/core/ext/transport/binder/wire_format/binder_android.cc \
-src/core/ext/transport/binder/wire_format/binder_android.h \
-src/core/ext/transport/binder/wire_format/binder_constants.cc \
-src/core/ext/transport/binder/wire_format/binder_constants.h \
-src/core/ext/transport/binder/wire_format/transaction.cc \
-src/core/ext/transport/binder/wire_format/transaction.h \
-src/core/ext/transport/binder/wire_format/wire_reader.h \
-src/core/ext/transport/binder/wire_format/wire_reader_impl.cc \
-src/core/ext/transport/binder/wire_format/wire_reader_impl.h \
-src/core/ext/transport/binder/wire_format/wire_writer.cc \
-src/core/ext/transport/binder/wire_format/wire_writer.h \
 src/core/ext/transport/chttp2/README.md \
 src/core/ext/transport/chttp2/alpn/alpn.cc \
 src/core/ext/transport/chttp2/alpn/alpn.h \
